refactor: EquipmentRentalsDto 하나의 파일로 리팩토링
