--- src/main/java/com/girigiri/kwrental/rental/repository/dto/EquipmentRentalDto.java
@@ -1,36 +1,0 @@
-package com.girigiri.kwrental.rental.repository.dto;
-
-import lombok.Getter;
-
-import java.time.LocalDate;
-import java.util.Collection;
-import java.util.HashSet;
-import java.util.Objects;
-import java.util.Set;
-
-@Getter
-public class EquipmentRentalDto {
-
-    private final LocalDate startDate;
-    private final LocalDate endDate;
-    private final Set<RentalSpecDto> rentalSpecs;
-
-    public EquipmentRentalDto(final LocalDate startDate, final LocalDate endDate, final Collection<RentalSpecDto> rentalSpecs) {
-        this.startDate = startDate;
-        this.endDate = endDate;
-        this.rentalSpecs = new HashSet<>(rentalSpecs);
-    }
-
-    @Override
-    public boolean equals(final Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        final EquipmentRentalDto rentalDto = (EquipmentRentalDto) o;
-        return Objects.equals(startDate, rentalDto.startDate) && Objects.equals(endDate, rentalDto.endDate) && Objects.equals(rentalSpecs, rentalDto.rentalSpecs);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(startDate, endDate, rentalSpecs);
-    }
-}
